refactor: DexieRewriteRuleRepositoryをframeworks-and-drivers/persistenceに移動
user-story-001の分類Bファイル移動タスクとして、
DexieRewriteRuleRepository.tsをClean Architectureの
正しい層(frameworks-and-drivers)に移動

- src/infrastructure/persistence/indexeddb/ から
  src/frameworks-and-drivers/persistence/ へ移動
- 関連するテストファイルも同様に移動
- import文を全て新パスに更新
