+++ host-frontend-root/frontend-src-root/src/frameworks-and-drivers/persistence/DexieRewriteRuleRepository.ts
