+++ host-frontend-root/frontend-src-root/tests/unit/frameworks-and-drivers/persistence/DexieRewriteRuleRepository/getById/normal-cases.test.ts
@@ -4,7 +4,7 @@ import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 
 import { RewriteRule } from 'src/enterprise-business-rules/entities/RewriteRule/RewriteRule';
 import { dexieDatabase } from 'src/infrastructure/persistence/indexeddb/DexieDatabase';
-import { DexieRewriteRuleRepository } from 'src/infrastructure/persistence/indexeddb/DexieRewriteRuleRepository';
+import { DexieRewriteRuleRepository } from 'src/frameworks-and-drivers/persistence/DexieRewriteRuleRepository';
 
 /**
  * 1. 指定されたIDのルールが存在する場合、そのRewriteRuleインスタンスを返却
